+++ lib/rails-creds.rb
@@ -1,3 +1,1 @@
-# rubocop:disable Naming/FileName
-# rubocop:enable Naming/FileName
-require 'creds'
+require "creds"